Implement md5 hashing for file identification.

+++ src/headers/api/AudioData.h
@@ -4,7 +4,7 @@
 
 namespace iamaprogrammer {
   struct AudioData {
-    long long frames;
+    size_t frames;
     int sampleRate;
     int channels;
   };
